Fixes MRTK3's `Tap to Place` solver to work with any hand and interactor (#11545)
## Overview
Fixes MRTK3's `Tap to Place` solver to work with any hand and any interactor (even speech). This change, on `StartPlacement`, queries the `XRInteractionManager` for all registered interactors, and then registers for the interactors' select events.  This is an alternative to requiring the developer to specify particular actions. 

Note, this change also removes the `StatefulInteractable` requirement.  The consumer of `TapToPlace` is now required to determine when `StartPlacement` is invoked. This change is meant to provide extensibility for future scenarios which may not require or use `StatefulInteractables`.

This also adds Unit Tests to validate `TapToPlace`.

## Changes
- Fixes: https://github.com/microsoft/MixedRealityToolkit-Unity/issues/11527