Add Scale and Rotate Cursor to MRTK3 (#11603)
Adds a SpatialManipulationReticle component, which allows a spatial manipulation cursor prefab to be designated as scale or rotate. Makes changes to MRTKRayReticleVisual to use this component to appropriately rotate the cursor, if it exists on the custom reticle. Cursor position and rotation is updated every frame.

Changes to MRTKRayInteractor allow objects to continue to be hovered as they are selected and moved.