The intent of this change is to make attaching a custom reticle easier in MRTK. The custom reticle is now attached to the base recticle's root transform, removing the need for the custom reticle to position itself.
Also, reverting changes made to `RecticleMagnetism`, this now is using an interface again for setting progress, instead of requiring a RingReticle component.  Instead of using the old `IVariableReticle` its not `IVariableProgressReticle` to better indicate the propose of the interface.

I also rename the newer `IVariableReticle`  to `RecticleVisual` to better indicate its purpose.